Decoupled qs_index from the array index of questions.
Decoupled qs_index from the array index of questions being inserted into
the database.  Array entries are not guaranteed to be sequential, but we
would like them sequential being inserted into the database.  Looping
thru the array expecting a specific index was broken when users entered
an deleted rows via the application.